build to .exe v1.1
